meet: move config schema → skills/meet-join/config-schema.ts
