+++ skills/meet-join/config-schema.ts
