+++ skills/meet-join/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@vellumai/meet-skill",
+  "version": "0.0.1",
+  "private": true,
+  "license": "MIT",
+  "type": "module",
+  "dependencies": {
+    "zod": "4.3.6"
+  },
+  "devDependencies": {
+    "@types/bun": "1.2.4",
+    "typescript": "5.7.3"
+  }
+}
